perf/core: Implement workaround to fix RCU stall issue when using perf record
This patch is a workaround for now. As the root cause has not yet been
identified, we will continue to search for the root cause and resolve it
in the future.

When using the 'perf record' command, an RCU (Read-Copy-Update) stall
issue is encountered.
The error log is shown as below:

[ 52.012118  ] rcu: INFO: rcu_preempt detected stalls on CPUs/tasks:
[ 52.030451  ] rcu: 3-...0: (2 ticks this GP)
idle=366c/1/0x4000000000000002 softirq=378/378 fqs=2615
[ 52.057510  ] (detected by 1, t=5259 jiffies, g=1437, q=13 ncpus=4)
[ 52.076009  ] Task dump for CPU 3:
[ 52.085725  ] task:perf-exec state:R running task stack:0
pid:96 ppid:93 flags:0x0000000a
[ 52.115443  ] Call Trace:
[ 52.122860  ] [] __schedule+0x4cc/0x554

We have found that RCU stall is due to an excessive number of perf
interrupts. Although the perf process itself seems to be functioning
properly, but there is a potential issue with the Linux timer subsystem
that causes the perf period to remain consistently at 1. As a result,
overflow interrupts are triggered on every event, resulting in a large
number of perf interrupts.

We have discovered that modifying the parameters passed into
perf_adjust_period() to calculate the difference of HPMcounter every
time, instead of relying on the last period value, can effectively
change the value of the period and prevent from the occurrence of RCU
stall issues.

Note again that this is only a workaround, not a final solution. This
patch should be reverted once the root cause is found.

Signed-off-by: Locus Wei-Han Chen <locus84@andestech.com>