andes: ftsdc010: Add another revision number
According to Kavalan/James <jamesykm@andestech.com>,
the revision number for future Andes SDC IP will be changed to 32'h0003_0108.
Therefore, for backward compatibility, we add another revision number
to accommodate this change.

Based on the FPGA bitmap feedback from Condor, the MMC controller has
a revision ID of 0x00030106. Therefore, we have added support for this
revision of the MMC controller.

Signed-off-by: Leo Yu-Chi Liang <ycliang@andestech.com>
Signed-off-by: CL Wang  <cl634@andestech.com>