riscv: dts: andes: add dma global pool in the reserved memory
In the Andes platform, the PPMA is used to set up noncacheable regions
for noncoherent devices. Due to the limited number of PPMA entries,
the number of noncacheable regions that can be set is also limited.

By using a DMA global pool, a large noncacheable region consisting
of multiple pages can be configured with just one PPMA entry.

The DMA global pool is also suitable for environments that
use MSB and Svpbmt to set noncacheable regions.

Additionally, the DMA global pool has been added to reserved memory,
providing a total of 512 pages in size.

Signed-off-by: Charles Ci-Jyun Wu <dminus@andestech.com>