soc: andes: dcause: Add the "print_detailed_cause" method
When multiple events cause a trap to be taken with the same scause value,
we need to know the sdcause value that records the detailed cause.
The "print_detailed_cause" method is utilised to print out the detailed
cause when the trap occurs.

Reformed from the following patches on ast-v5_4_0-branch:
- (825484e69e) drivers: soc: andes: Fix an error for print_detailed_cause() on imprecise exception (#101)

Signed-off-by: Ben Zong-You Xie <ben717@andestech.com>

To print error messages for imprecise exception, extract the field sdcause
in the CSR sdcause instead of reading the whole register.

Signed-off-by: Ben Zong-You Xie <ben717@andestech.com>

Reviewed-on: https://gitea.andestech.com/RD-SW/linux/pulls/101
Reviewed-by: Tim Shih-Ting OuYang <tim609@andestech.com>
Reviewed-by: Leo Yu-Chi Liang <ycliang@andestech.com>
Reviewed-by: Dylan Dai-Rong Jhong <dylan@andestech.com>
Co-authored-by: Ben Zong-You Xie <ben717@andestech.com>
Co-committed-by: Ben Zong-You Xie <ben717@andestech.com>