riscv: andes: fix MSB setting failure for huge page in 64-bit system
When using the MSB mechanism to set uncacheable mappings on
a 64-bit system, calling ioremap_wc() with a 2MiB huge page
to set the MSB fails.

This issue arises because the original MSB setting procedure involves
calling ioremap_wc() to set _PAGE_ANDES25_NOCACHE in pgprot, allowing
pfn_pte() to configure the MSB in the PFN of a 4KiB page and subsequently
create the PTE. This process overlooks scenarios where different
page sizes are used for mapping operations.

ref : (310f541a027b) riscv: Enable HAVE_ARCH_HUGE_VMAP for 64BIT

Signed-off-by: charles <dminus@andestech.com>