smu: andes: kernel/workqueue: Fix delay wakeup by setting default workqueue scope affinity to NUMA (#164)
We git bisected and found that the commit 63c5484e7495
("workqueue: Add multiple affinity scopes and interface to select them")
would cause the system to wakeup after, but not as soon as, wakeup event is triggered.

Instead of reverting the commit, changing the default workqueue attribute alone
to NUMA like before could solve this issue.

Fixes: 63c5484e7495 ("workqueue: Add multiple affinity scopes and interface to select them")
Signed-off-by: Leo Yu-Chi Liang <ycliang@andestech.com>