spi: andes: support spi peripherals (#155)
Previously, the atcspi200 driver only supported SPI NOR flash.
However, due to ticket 26070, a customer requested that the atcspi200 driver
support SPI slave devices.Therefore, this functionality has been added.

Since we do not have the customer's slave device, the testing was
conducted using an SPI NOR flash as a substitute, and we used the read
ID command to verify if it met expectations.

Signed-off-by: Locus Wei-Han Chen <locus84@andestech.com>
Reviewed-on: https://gitea.andestech.com/RD-SW/linux/pulls/155
Reviewed-by: Leo Yu-Chi Liang <ycliang@andestech.com>
Reviewed-by: randolph <randolph@andestech.com>
Reviewed-by: Tim Shih-Ting OuYang <tim609@andestech.com>
Reviewed-by: CL Chin-Long Wang <cl634@andestech.com>
Reviewed-by: Charles Ci-Jyun Wu <dminus@andestech.com>
Co-authored-by: Locus Wei-Han Chen <locus84@andestech.com>
Co-committed-by: Locus Wei-Han Chen <locus84@andestech.com>