mmc: andes: ftsdc010: Remove offensive words from code and set dma_filter to NULL in DMA channel requests (#165)
In addition to the native definitions in the Linux kernel, remove the word
"slave" from variable names, function names, define, and comments.

The ftsdc010 driver obtains a DMA channel through dma_request_channel.
Since there is no need to restrict the allocated DMA channel, the
filter_para and dma_filter parameters are removed to align with the
current driver implementation.

Signed-off-by: CL Wang <cl634@andestech.com>
Reviewed-on: https://gitea.andestech.com/RD-SW/linux/pulls/165
Reviewed-by: Tim Shih-Ting OuYang <tim609@andestech.com>
Reviewed-by: randolph <randolph@andestech.com>
Reviewed-by: Leo Yu-Chi Liang <ycliang@andestech.com>
Co-authored-by: CL Wang <cl634@andestech.com>
Co-committed-by: CL Wang <cl634@andestech.com>