watchdog: andes: atcwdt200: Use div64_s64 for 64-bit division to avoid link error with 32-bit toolchain (#246)
The following code causes a link error when using a 32-bit toolchain due
to an undefined reference to __divdi3

rest_time_ms =
        timeout * 1000LL - ((1000LL << int_index) / drv_data->clk_freq);

This occurs because 64-bit division (/) in a 32-bit environment requires
the __divdi3 function, which is not available. To resolve this, replace
the division operation with div64_s64 to handle 64-bit division safely
in a 32-bit toolchain environment.

Reviewed-on: https://gitea.andestech.com/RD-SW/linux/pulls/246
Reviewed-by: randolph <randolph@andestech.com>
Reviewed-by: Locus Wei-Han Chen <locus84@andestech.com>
Reviewed-by: Mina Hui-Min Chou <minachou@andestech.com>
Co-authored-by: CL Wang <cl634@andestech.com>
Co-committed-by: CL Wang <cl634@andestech.com>