selftest: ftrace: Prevent attaching handlers and nested functions to kprobe events (#279)
Attaching a handler or nested functions to kprobe events is not allowed.
Doing so may cause the probe to recursively trigger itself,
leading to repeated invocation of the registered function.
This results in stack overflows and ultimately crashes the kernel.

To avoid such issues, ensure that probe targets do not include
the handler or any nested function that might be part of the probe path.

Signed-off-by: Cynthia Huang <cynthia@andestech.com>
Reviewed-on: https://gitea.andestech.com/RD-SW/linux/pulls/279
Reviewed-by: Mina Hui-Min Chou <minachou@andestech.com>
Reviewed-by: Tim Shih-Ting OuYang <tim609@andestech.com>
Co-authored-by: Cynthia Huang <cynthia@andestech.com>
Co-committed-by: Cynthia Huang <cynthia@andestech.com>